requirements.txt
pandas~=0.24.2
configparser~=4.0.2
pyvcf